예제코드 - Factory 파일 위치 수정
