+++ src/FactoryMethod/framework/Product.java
@@ -1,0 +1,5 @@
+package FactoryMethod.framework;
+// FactoryMethod.framework 는 제품에 대해서 표현한 패키지, 사용 생성 이런것
+public abstract  class Product {
+    public abstract void use();
+}
