+++ src/FactoryMethod/idcard/IDCardFactory.java
@@ -1,30 +1,20 @@
-package idcard;
-import framework.*;
+package FactoryMethod.idcard;
+import FactoryMethod.framework.*;
 import java.util.*;
 
 public class IDCardFactory extends Factory {
-    private HashMap db = new HashMap();
-    private int serial=100;
-
     private List owners = new ArrayList();
 
     @Override
-    // synchronized -
     protected Product createProduct(String owner) {
-        return new IDCard(owner, serial++);
+        return new IDCard(owner);
     }
 
     @Override
     protected void registerProduct(Product product) {
-        IDCard card = (IDCard) product;
-        db.put(new Integer(card.getSerial()), card.getOwner());
-
         owners.add(((IDCard)product).getOwner());
     }
     public List getOwners() {
         return owners;
-    }
-    public HashMap getDb() {
-        return db;
     }
 }
